Reorganize Class directory structure and placement
